experiments finally worked
the "other tutorials" folder contains a couple of client/server stacks
that did work for me, at least not for what I was trying to do. "Chat
server" worked, so I made a copy (chat server 2) and modified it with
the "myClient" substack which calls on "startServer" to begin the python
server and displays the data when the server sends it.